Create EOIminor
This is a small project from EOI subject by our team members.
It is a To-Do list in which we can create task, delete tasks , date due , mark them as completed, etc. 
We have used simple concepts of cpp programming for this which includes concepts like functions, array, switch case menu, etc.